add corporate model and corporate controllers

+++ controllers/employee-controller.js
@@ -1,9 +1,13 @@
 const Employee = require("../models/employee-model");
+const Corporate = require("../models/corporate-model");
 const { AppError } = require("../utils/app-error");
+const { ObjectId } = require("mongoose").Types;
 
 const getAllEmployees = async (request, response, next) => {
   try {
-    const employees = await Employee.find({});
+    const employees = await Employee.find({})
+      .select("firstName lastName corporate")
+      .populate("corporate");
 
     response.status(200).json({
       status: "success",
@@ -17,7 +21,7 @@ const getAllEmployees = async (request, response, next) => {
 const getOneEmployee = async (request, response, next) => {
   try {
     const { id } = request.params;
-    const employee = await Employee.findById({ _id: id });
+    const employee = await Employee.findById({ _id: id }).populate("corporate");
 
     response.status(200).json({
       status: "success",
@@ -30,9 +34,18 @@ const getOneEmployee = async (request, response, next) => {
 
 const addEmployee = async (request, response, next) => {
   try {
-    // await Employee.exists({ phoneNumber: request.body.phoneNumber });
+    if (!isValidObjectId(request.body.corporate)) {
+      return next(new AppError(400, "invalid corporate id"));
+    }
+
+    const corporate = await Corporate.findById(request.body.corporate);
+    if (!corporate) {
+      return next(new AppError(404, "corporate id not-found "));
+    }
 
     const employee = new Employee(request.body);
+
+    await employee.save();
     console.info(`employee (id:${employee._id}}) added successfully.`);
 
     response.status(200).json({
@@ -40,6 +53,7 @@ const addEmployee = async (request, response, next) => {
       data: { employee },
     });
   } catch (err) {
+    console.log(err);
     next(new AppError(500, err.message));
   }
 };
@@ -73,81 +87,47 @@ const showEmployeesData = async (request, response, next) => {
 };
 
 const updateEmployeeData = async (request, response, next) => {
-  const { id } = request.params;
-  const data = request.body;
-  const employeeWithNewInfo = await Employee.findByIdAndUpdate(id, data, {
-    new: true,
-  });
-  if (!employeeWithNewInfo) {
-    next(new AppError(404, "Employee not found"));
-  }
+  try {
+    const employeeWithNewInfo = await Employee.findByIdAndUpdate(
+      request.params.id,
+      request.body,
+      {
+        new: true,
+      }
+    );
+    if (!employeeWithNewInfo) {
+      next(new AppError(404, "Employee not found"));
+    }
 
-  response.status(200).json({ status: "success", data: employeeWithNewInfo });
+    response.status(200).json({ status: "success", data: employeeWithNewInfo });
+  } catch (error) {
+    next(new AppError(500, error.message));
+  }
 };
 
 const removeEmplpyeeData = async (request, response, next) => {
   const { id } = request.params;
-  const employeeWithNewInfo = await Employee.findByIdAndDelete(id);
-  if (!employeeWithNewInfo) {
+  if (!isValidObjectId(id)) {
+    return next(new AppError(400, "invalid employee id"));
+  }
+
+  const employee = await Employee.findByIdAndDelete(id);
+  if (!employee) {
     next(new AppError(404, "Employee not fund"));
   }
 
   response.status(204).json({ status: "success", data: { message: "employee removed" } });
 };
 
-const findEmployeesByCity = async (request, response, next) => {
-  const { city } = request.params;
-  const employeesInOneCity = await Employee.find({ city });
-
-  response.status(200).json({ status: "success", data: employeesInOneCity });
-};
-
-const findEmployeesByage = async (request, response, next) => {
-  const { age: targetAge } = request.params;
-  const employees = await Employee.find({});
-
-  const employeesOlderThanAge = employees.filter((employee) => {
-    const age = new Date().getFullYear() - new Date(employee.birthDate).getFullYear();
-    if (age < targetAge) {
-      return false;
+function isValidObjectId(id) {
+  if (ObjectId.isValid(id)) {
+    if (new ObjectId(id).toString() === id) {
+      return true;
     }
-    return true;
-  });
-
-  response.status(200).json({ status: "success", data: employeesOlderThanAge });
-};
-
-const findEmployeesByPhoneCount = async (request, response, next) => {
-  const { count } = request.params;
-  const employeesWithOnePhoneNumber = await Employee.find({
-    phoneNumber: { $size: count },
-  });
-
-  response.status(200).json({ status: "success", data: employeesWithOnePhoneNumber });
-};
-
-const findWomenManager = async (request, response, next) => {
-  const employeesWomenAndManager = await Employee.find({
-    gender: "woman",
-    role: "manager",
-  });
-
-  response.status(200).json({ status: "success", data: employeesWomenAndManager });
-};
-
-const sortEmployeeMenByAge = async (request, response, next) => {
-  const employeesMen = await Employee.find({ gender: "man" });
-
-  const employeesMenWithAge = employeesMen.map((employee) => {
-    const age = new Date().getFullYear() - new Date(employee.birthDate).getFullYear();
-
-    return { ...employee.toObject(), age };
-  });
-
-  const sortEmployeesWithAge = employeesMenWithAge.sort((a, b) => b.age - a.age);
-
-  response.status(200).json({ status: "success", data: sortEmployeesWithAge });
-};
+    return false;
+  }
+  return false;
+}
 
 module.exports = {
   getAllEmployees,
@@ -157,9 +137,4 @@ module.exports = {
   showEmployeesData,
   updateEmployeeData,
   removeEmplpyeeData,
-  findEmployeesByCity,
-  findEmployeesByage,
-  findEmployeesByPhoneCount,
-  findWomenManager,
-  sortEmployeeMenByAge,
 };
